+++ api/server.go
@@ -7,9 +7,17 @@ import (
 	"github.com/dogmatiq/interopspec/configspec"
 )
 
-// NewServer returns a configspec.ConfigAPIServer for the given applications.
-func NewServer(apps ...configkit.Application) configspec.ConfigAPIServer {
-	s := &server{}
+// Server is an implementation of configspec.ConfigAPIServer.
+type Server struct {
+	response configspec.ListApplicationsResponse
+}
+
+var _ configspec.ConfigAPIServer = (*Server)(nil)
+
+// NewServer returns an API server that servers the configuration of the given
+// applications.
+func NewServer(apps ...configkit.Application) *Server {
+	s := &Server{}
 
 	for _, in := range apps {
 		out, err := marshalApplication(in)
@@ -17,8 +25,8 @@ func NewServer(apps ...configkit.Application) configspec.ConfigAPIServer {
 			panic(err)
 		}
 
-		s.ListApplicationsResponse.Applications = append(
-			s.ListApplicationsResponse.Applications,
+		s.response.Applications = append(
+			s.response.Applications,
 			out,
 		)
 	}
@@ -26,14 +34,10 @@ func NewServer(apps ...configkit.Application) configspec.ConfigAPIServer {
 	return s
 }
 
-type server struct {
-	configspec.ListApplicationsResponse
-}
-
 // ListApplications returns the full configuration of all applications.
-func (s *server) ListApplications(
+func (s *Server) ListApplications(
 	context.Context,
 	*configspec.ListApplicationsRequest,
 ) (*configspec.ListApplicationsResponse, error) {
-	return &s.ListApplicationsResponse, nil
+	return &s.response, nil
 }